+++ lib/lhm/entangler.rb
@@ -3,6 +3,7 @@
 
 require 'lhm/command'
 require 'lhm/sql_helper'
+require 'lhm/sql_retry'
 
 module Lhm
   class Entangler
@@ -11,9 +12,6 @@ module Lhm
 
     attr_reader :connection
 
-    LOCK_WAIT_RETRIES = 10
-    RETRY_WAIT = 1
-
     # Creates entanglement between two tables. All creates, updates and deletes
     # to origin will be repeated on the destination table.
     def initialize(migration, connection = nil, options = {})
@@ -21,8 +19,12 @@ module Lhm
       @origin = migration.origin
       @destination = migration.destination
       @connection = connection
-      @max_retries = options[:lock_wait_retries] || LOCK_WAIT_RETRIES
-      @sleep_duration = options[:retry_wait] || RETRY_WAIT
+      @retry_helper = SqlRetry.new(
+        @connection,
+        {
+          log_prefix: "Entangler"
+        }.merge!(options.fetch(:retriable, {}))
+      )
     end
 
     def entangle
@@ -88,13 +90,17 @@ module Lhm
 
     def before
       entangle.each do |stmt|
-        with_retry { @connection.execute(tagged(stmt)) }
+        @retry_helper.with_retries do |retriable_connection|
+          retriable_connection.execute(stmt)
+        end
       end
     end
 
     def after
       untangle.each do |stmt|
-        with_retry { @connection.execute(tagged(stmt)) }
+        @retry_helper.with_retries do |retriable_connection|
+          retriable_connection.execute(stmt)
+        end
       end
     end
 
@@ -106,22 +112,6 @@ module Lhm
 
     def strip(sql)
       sql.strip.gsub(/\n */, "\n")
-    end
-
-    def with_retry
-      begin
-        retries ||= 0
-        yield
-      rescue StandardError => e
-        if e.message =~ /Lock wait timeout exceeded/ && retries < @max_retries
-          retries += 1
-          Lhm.logger.info("#{e} - retrying #{retries} time(s)")
-          Kernel.sleep @sleep_duration
-          retry
-        else
-          raise e
-        end
-      end
     end
   end
 end