+++ spec/integration/atomic_switcher_spec.rb
@@ -18,7 +18,8 @@ describe Lhm::AtomicSwitcher do
       @origin      = table_create('origin')
       @destination = table_create('destination')
       @migration   = Lhm::Migration.new(@origin, @destination)
-      Lhm.logger = Logger.new('/dev/null')
+      @logs = StringIO.new
+      Lhm.logger = Logger.new(@logs)
       @connection.execute('SET GLOBAL innodb_lock_wait_timeout=3')
       @connection.execute('SET GLOBAL lock_wait_timeout=3')
     end
@@ -27,38 +28,46 @@ describe Lhm::AtomicSwitcher do
       Thread.abort_on_exception = false
     end
 
-    it 'should retry on lock wait timeouts' do
-      skip 'This spec only works with mysql2' unless defined? Mysql2
+    it 'should retry and log on lock wait timeouts' do
+      connection = mock()
+      connection.stubs(:data_source_exists?).returns(true)
+      connection.stubs(:execute).raises(ActiveRecord::StatementInvalid, 'Lock wait timeout exceeded; try restarting transaction.').then.returns(true)
 
-        connection = mock()
-        connection.stubs(:data_source_exists?).returns(true)
-        connection.stubs(:execute).raises(ActiveRecord::StatementInvalid, 'Lock wait timeout exceeded; try restarting transaction.').then.returns(true)
+      switcher = Lhm::AtomicSwitcher.new(@migration, connection, retriable: {base_interval: 0})
 
-        switcher = Lhm::AtomicSwitcher.new(@migration, connection)
-        switcher.retry_sleep_time = 0
+      assert switcher.run
 
-        assert switcher.run
+      log_messages = @logs.string.split("\n")
+      assert_equal(2, log_messages.length)
+      assert log_messages[0].include? "Starting run of class=Lhm::AtomicSwitcher"
+      assert log_messages[1].include? "[AtomicSwitcher] ActiveRecord::StatementInvalid: 'Lock wait timeout exceeded; try restarting transaction.' - 1 tries"
     end
 
-    it 'should give up on lock wait timeouts after MAX_RETRIES' do
-      skip 'This spec only works with mysql2' unless defined? Mysql2
-        connection = mock()
-        connection.stubs(:data_source_exists?).returns(true)
-        connection.stubs(:execute).twice.raises(ActiveRecord::StatementInvalid, 'Lock wait timeout exceeded; try restarting transaction.')
+    it 'should give up on lock wait timeouts after a configured number of tries' do
+      connection = mock()
+      connection.stubs(:data_source_exists?).returns(true)
+      connection.stubs(:execute).twice.raises(ActiveRecord::StatementInvalid, 'Lock wait timeout exceeded; try restarting transaction.')
 
-        switcher = Lhm::AtomicSwitcher.new(@migration, connection)
-        switcher.max_retries = 2
-        switcher.retry_sleep_time = 0
+      switcher = Lhm::AtomicSwitcher.new(@migration, connection, retriable: {tries: 2, base_interval: 0})
 
-        assert_raises(ActiveRecord::StatementInvalid) { switcher.run }
+      assert_raises(ActiveRecord::StatementInvalid) { switcher.run }
     end
 
     it 'should raise on non lock wait timeout exceptions' do
       switcher = Lhm::AtomicSwitcher.new(@migration, connection)
-      switcher.send :define_singleton_method, :statements do
-        ['SELECT', '*', 'FROM', 'nonexistent']
+      switcher.send :define_singleton_method, :atomic_switch do
+        'SELECT * FROM nonexistent'
       end
       -> { switcher.run }.must_raise(ActiveRecord::StatementInvalid)
+    end
+
+    it "should raise when destination doesn't exist" do
+      connection = mock()
+      connection.stubs(:data_source_exists?).returns(false)
+
+      switcher = Lhm::AtomicSwitcher.new(@migration, connection)
+
+      assert_raises(Lhm::Error) { switcher.run }
     end
 
     it 'rename origin to archive' do