+++ spec/test_helper.rb
@@ -1,10 +1,15 @@
 # Copyright (c) 2011 - 2013, SoundCloud Ltd., Rany Keddo, Tobias Bielohlawek, Tobias
 # Schmidt
 
+if ENV['COV']
+  require 'simplecov'
+  SimpleCov.start
+end
+
 require 'minitest/autorun'
 require 'minitest/spec'
 require 'minitest/mock'
-require 'mocha/mini_test'
+require 'mocha/minitest'
 require 'pathname'
 require 'lhm'
 
@@ -13,11 +18,7 @@ $spec = $project.join('spec')
 $fixtures = $spec.join('fixtures')
 
 require 'active_record'
-begin
-  require 'mysql2'
-rescue LoadError
-  require 'mysql'
-end
+require 'mysql2'
 
 logger = Logger.new STDOUT
 logger.level = Logger::WARN