+++ spec/unit/throttler/slave_lag_spec.rb
@@ -36,9 +36,10 @@ describe Lhm::Throttler::Slave do
   include UnitHelper
 
   before :each do
-    def get_config
-      lambda { {'username' => 'user', 'password' => 'pw', 'database' => 'db'} }
-    end
+    @logs = StringIO.new
+    Lhm.logger = Logger.new(@logs)
+
+    @dummy_mysql_client_config = lambda { {'username' => 'user', 'password' => 'pw', 'database' => 'db'} }
   end
 
   describe "#client" do
@@ -52,24 +53,36 @@ describe Lhm::Throttler::Slave do
 
     describe 'on connection error' do
       it 'logs and returns nil' do
-        test_client = lambda { |config|
-          TestMysql2Client.new(config)
-        }
-        Mysql2::Client.stubs(:new).returns(test_client) do
-          assert_send([Lhm.logger, :info, "Error connecting to slave: connection error"])
-          assert_nil(Lhm::Throttler::Slave.new('slave', lambda { {} }).connection)
-        end
+        assert_nil(Lhm::Throttler::Slave.new('slave', @dummy_mysql_client_config).connection)
+
+        log_messages = @logs.string.lines
+        assert_equal(2, log_messages.length)
+        assert log_messages[0].include? "Connecting to slave on database: db"
+        assert log_messages[1].include? "Error connecting to slave: Unknown MySQL server host 'slave'"
       end
     end
 
     describe 'with proper config' do
       it "creates a new Mysql2::Client" do
-        client_assertion = lambda { |config|
-          assert_equal(config, {'username' => 'user', 'password' => 'pw', 'database' => 'db', 'host' => 'slave'})
-        }
-        Mysql2::Client.stubs(:new).returns(client_assertion) do
-          Lhm::Throttler::Slave.new('slave', get_config)
-        end
+        expected_config = {username: 'user', password: 'pw', database: 'db', host: 'slave'}
+        Mysql2::Client.stubs(:new).with(expected_config).returns(mock())
+
+        assert Lhm::Throttler::Slave.new('slave', @dummy_mysql_client_config).connection
+      end
+    end
+
+    describe 'with active record config' do
+      it 'logs and creates client' do
+        active_record_config = {username: 'user', password: 'pw', database: 'db'}
+        ActiveRecord::Base.stubs(:connection_pool).returns(stub(spec: stub(config: active_record_config)))
+
+        Mysql2::Client.stubs(:new).returns(mock())
+
+        assert Lhm::Throttler::Slave.new('slave').connection
+
+        log_messages = @logs.string.lines
+        assert_equal(1, log_messages.length)
+        assert log_messages[0].include? "Connecting to slave on database: db"
       end
     end
   end
@@ -86,7 +99,7 @@ describe Lhm::Throttler::Slave do
         end
       end
 
-      @slave = Lhm::Throttler::Slave.new('slave', get_config)
+      @slave = Lhm::Throttler::Slave.new('slave', @dummy_mysql_client_config)
       @slave.instance_variable_set(:@connection, Connection)
 
       class StoppedConnection
@@ -95,7 +108,7 @@ describe Lhm::Throttler::Slave do
         end
       end
 
-      @stopped_slave = Lhm::Throttler::Slave.new('stopped_slave', get_config)
+      @stopped_slave = Lhm::Throttler::Slave.new('stopped_slave', @dummy_mysql_client_config)
       @stopped_slave.instance_variable_set(:@connection, StoppedConnection)
     end
 
@@ -124,7 +137,7 @@ describe Lhm::Throttler::Slave do
         Lhm::Throttler::Slave.any_instance.stubs(:client).returns(client)
         Lhm::Throttler::Slave.any_instance.stubs(:config).returns([])
 
-        slave = Lhm::Throttler::Slave.new('slave', get_config)
+        slave = Lhm::Throttler::Slave.new('slave', @dummy_mysql_client_config)
         assert_send([Lhm.logger, :info, "Unable to connect and/or query slave: error"])
         assert_equal(0, slave.lag)
       end
@@ -210,7 +223,7 @@ describe Lhm::Throttler::SlaveLag do
         client.stubs(:query).raises(Mysql2::Error, "Can't connect to MySQL server")
         Lhm::Throttler::Slave.any_instance.stubs(:client).returns(client)
 
-        Lhm::Throttler::Slave.any_instance.stubs(:config).returns([])
+        Lhm::Throttler::Slave.any_instance.stubs(:prepare_connection_config).returns([])
         Lhm::Throttler::Slave.any_instance.stubs(:slave_hosts).returns(['1.1.1.2'])
         @throttler.stubs(:master_slave_hosts).returns(['1.1.1.1'])
 
@@ -237,7 +250,7 @@ describe Lhm::Throttler::SlaveLag do
         class TestSlave
           attr_reader :host, :connection
 
-          def initialize(host, get_config)
+          def initialize(host, _)
             @host = host
             @connection = 'conn' if @host
           end